virus ATP changes

+++ classes/ATP.py
@@ -25,7 +25,7 @@ class ATP(pygame.sprite.Sprite):
 
 	def addVirusATP(self, wave):
 		#FUNCTION: 100log(x)+100
-		self.currentVirusATP += 100*math.log(wave)+100
+		self.currentVirusATP += int(100*math.log(wave))+100
 		print self.currentVirusATP
 
 	def update(self):
